Don't show notification when playing is paused
When MPD is paused, it will write that fact to the state file. However,
currently we don't evaluate the state and so we send a notification as
if a song is being played.
Fix this issue by evaluating the daemon's state and bailing out if it
indicates that playing is paused.