Don't show repeated notifications
MPD updates the state file not only when a new song plays but also for
other reasons, such as to update the time a song is played at. From our
end, this can mean repeated notifications for the very same song, which
is something we are not interested in.
To fix this problem, remember what was playing previously and only show
a notification if it changed.